+++ openbrokerapi/__init__.py
@@ -1,0 +1,2 @@
+
+# TODO: Do not require deprecated fields if they are available in context